add BitVector class
