+++ src/core/Makefile.am
@@ -320,6 +320,7 @@ HEADERS_COMMON                             = \
     coap/coap.hpp                            \
     coap/coap_message.hpp                    \
     coap/coap_secure.hpp                     \
+    common/bit_vector.hpp                    \
     common/clearable.hpp                     \
     common/code_utils.hpp                    \
     common/crc16.hpp                         \
